Add static sheet map
